backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiomysql==0.2.0
aiosignal==1.3.2
alembic==1.13.3
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.0
arrow==1.3.0
attrs==25.3.0
azure-core==1.34.0
bcrypt==4.1.2
binaryornot==0.4.4
bleach==6.2.0
blinker==1.8.2
botbuilder-core==4.17.0
botbuilder-integration-aiohttp==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.6
colorama==0.4.6
cookiecutter==2.6.0
cryptography==41.0.7
Deprecated==1.2.18
dnspython==2.7.0
EasyProcess==1.1
ecdsa==0.19.1
email-validator==2.1.0
entrypoint2==1.1
fastapi==0.104.1
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
frozenlist==1.7.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
isodate==0.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==1.4.2
limits==5.1.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
MouseInfo==0.1.3
msal==1.32.3
msrest==0.7.1
mss==9.0.1
multidict==6.4.4
mysql-connector-python==8.2.0
oauthlib==3.2.2
ordered-set==4.1.0
packaging==25.0
passlib==1.7.4
pdfkit==1.0.0
Pillow==10.1.0
propcache==0.3.2
protobuf==4.21.12
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyodbc==4.0.39
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.4
pytweening==1.0.7
PyYAML==6.0.2
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.24.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.2.3
wrapt==1.17.2
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
yarl==1.20.1
